fix: change fetch function to correct place
